+++ base-spring-boot-starter/src/main/java/eu/fraho/spring/security_jwt/starter/SecurityJwtNoRefreshStoreAutoConfiguration.java
@@ -1,0 +1,24 @@
+/*
+ * MIT Licence
+ * Copyright (c) 2017 Simon Frankenberger
+ *
+ * Please see LICENCE.md for complete licence text.
+ */
+package eu.fraho.spring.security_jwt.starter;
+
+import eu.fraho.spring.security.base.service.NullTokenStore;
+import eu.fraho.spring.security.base.service.RefreshTokenStore;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@AutoConfigureAfter(SecurityJwtBaseAutoConfiguration.class)
+public class SecurityJwtNoRefreshStoreAutoConfiguration {
+    @Bean
+    @ConditionalOnMissingBean
+    public RefreshTokenStore refreshTokenStore() {
+        return new NullTokenStore();
+    }
+}